Moved edit-profile.js file under pages folder under App and using tailwind instead of .css file
